requirements.txt
asgiref==3.2.10
astroid==2.3.3
certifi==2020.6.20
colorama==0.4.1
Django==3.0.3
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.4.3
pytz==2020.1
sqlparse==0.3.1
wincertstore==0.2
wrapt==1.11.2
asgiref==3.2.10
astroid==2.3.3
certifi==2020.6.20
colorama==0.4.1
Django==3.0.3
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
pylint==2.4.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
wincertstore==0.2
wrapt==1.11.2
